Refactor style
