+++ client/src/style/fonts.rs
@@ -9,7 +9,7 @@ use crate::prelude::*;
 /// * `weight` - Font weight.
 /// * `srcs` - Source urls
 ///
-pub fn font(family: &str, local: &str, style: Style, weight: Weight, srcs: &[&str]) -> css::Style {
+pub fn create_font(family: &str, local: &str, style: Style, weight: Weight, srcs: &[&str]) -> css::Style {
 	let mut ret = css::Style(Vec::new());
 
 	for &src in srcs {
@@ -28,15 +28,15 @@ pub fn font(family: &str, local: &str, style: Style, weight: Weight, srcs: &[&st
 }
 
 pub fn fonts() -> css::Style {
-	font("Alfa Slab One", "Alfa-Slab-One", Style::Normal, Weight::Normal, &["https://fonts.gstatic.com/s/alfaslabone/v10/6NUQ8FmMKwSEKjnm5-4v-4Jh2dJhew.woff2"])
-	+ font("Degrassi", "Degrassi Regular", Style::Normal, Weight::Normal, &["../public/fonts/degrassi/degrassi.ttf"])
-	+ font("Space Mono", "Space Mono Regular", Style::Normal, Weight::Number(400), &[
+	create_font("Alfa Slab One", "Alfa-Slab-One", Style::Normal, Weight::Normal, &["https://fonts.gstatic.com/s/alfaslabone/v10/6NUQ8FmMKwSEKjnm5-4v-4Jh2dJhew.woff2"])
+	+ create_font("Degrassi", "Degrassi Regular", Style::Normal, Weight::Normal, &["../public/fonts/degrassi/degrassi.ttf"])
+	+ create_font("Space Mono", "Space Mono Regular", Style::Normal, Weight::Number(400), &[
 			"https://fonts.gstatic.com/s/spacemono/v6/i7dPIFZifjKcF5UAWdDRYE58RWq7.woff2",
 			"https://fonts.gstatic.com/s/spacemono/v6/i7dPIFZifjKcF5UAWdDRYE98RWq7.woff2",
 			"https://fonts.gstatic.com/s/spacemono/v6/i7dPIFZifjKcF5UAWdDRYEF8RQ.woff2",
 		]
 	)
-	+ font("Space Mono", "Space Mono Bold", Style::Normal, Weight::Number(700), &[
+	+ create_font("Space Mono", "Space Mono Bold", Style::Normal, Weight::Number(700), &[
 			"https://fonts.gstatic.com/s/spacemono/v6/i7dMIFZifjKcF5UAWdDRaPpZUFqaHjyV.woff2",
 			"https://fonts.gstatic.com/s/spacemono/v6/i7dMIFZifjKcF5UAWdDRaPpZUFuaHjyV.woff2",
 			"https://fonts.gstatic.com/s/spacemono/v6/i7dMIFZifjKcF5UAWdDRaPpZUFWaHg.woff2",
